requirements.txt
---use-pep517
numpy>=1.26.0  # Python 3.12 requires numpy ≥1.26.0[1][3][5]

# API clients
anthropic==0.18.1

# Web and networking
flask==2.3.3
werkzeug==3.0.1
gunicorn==21.2.0
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
backoff==2.2.1
python-dotenv==1.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-limiter==3.5.0

# Data processing
python-dateutil==2.8.2
numpy==1.24.4

# Utility
pillow==11.0.0

# Bluefin Exchange API clients
git+https://github.com/fireflyprotocol/bluefin-client-python-sui.git
git+https://github.com/fireflyprotocol/bluefin-v2-client-python.git

# New dependencies
httpx==0.26.0
websockets==12.0

# Security libraries
pyjwt==2.8.0
cryptography==42.0.0
